satgui/requirements.txt
Django==3.0.3
django-crispy-forms==1.9.0
django-formtools==2.2
djangorestframework==3.11.0